requirements.txt
openai
pandas
plotly
streamlit
sqlite-utils
python-dotenv
scipy